# services/core-schedule/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "core-schedule",
    "description": "Schedule service — university timetable, templates, overrides",
    "version": "0.1.0"
  },
  "paths": {
...
    "/teachers/{teacher_name}/schedule": {
      "get": {
        "tags": ["teachers"],
        "summary": "Teacher Schedule",
        "operationId": "teacher_schedule_teachers__teacher_name__schedule_get",
        "parameters": [
          {
            "name": "teacher_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Teacher Name"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date To"
            }
          },
          {
            "name": "semester_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Semester Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DaySchedule"
                  },
                  "title": "Response Teacher Schedule Teachers  Teacher Name  Schedule Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teachers/{teacher_name}/exams": {
      "get": {
        "tags": ["teachers"],
        "summary": "Teacher Exams",
        "operationId": "teacher_exams_teachers__teacher_name__exams_get",
        "parameters": [
          {
            "name": "teacher_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Teacher Name"
            }
          },
          {
            "name": "semester_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Semester Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionEventResponse"
                  },
                  "title": "Response Teacher Exams Teachers  Teacher Name  Exams Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teachers/{teacher_name}/template": {
      "get": {
        "tags": ["teachers"],
        "summary": "Teacher Template",
        "operationId": "teacher_template_teachers__teacher_name__template_get",
        "parameters": [
          {
            "name": "teacher_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Teacher Name"
            }
          },
          {
            "name": "semester_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Semester Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DaySchedule": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "day_of_week": {
            "type": "integer",
            "title": "Day Of Week"
          },
          "day_name": {
            "type": "string",
            "title": "Day Name"
          },
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/LessonSlot"
            },
            "type": "array",
            "title": "Lessons"
          }
        },
        "type": "object",
        "required": ["date", "day_of_week", "day_name", "lessons"],
        "title": "DaySchedule"
      },
      "DayTemplate": {
        "properties": {
          "day_of_week": {
            "type": "integer",
            "title": "Day Of Week"
          },
          "day_name": {
            "type": "string",
            "title": "Day Name"
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/TemplateSlotResponse"
            },
            "type": "array",
            "title": "Slots"
          }
        },
        "type": "object",
        "required": ["day_of_week", "day_name", "slots"],
        "title": "DayTemplate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LessonSlot": {
        "properties": {
          "slot_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Number"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "title": "End Time"
          },
          "subject_name": {
            "type": "string",
            "title": "Subject Name"
          },
          "lesson_type": {
            "type": "string",
            "title": "Lesson Type"
          },
          "teacher_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teacher Name"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "classroom_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classroom Name"
          },
          "is_override": {
            "type": "boolean",
            "title": "Is Override",
            "default": false
          },
          "override_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Comment"
          }
        },
        "type": "object",
        "required": [
          "slot_number",
          "start_time",
          "end_time",
          "subject_name",
          "lesson_type",
          "teacher_name",
          "group_name",
          "classroom_name"
        ],
        "title": "LessonSlot"
      },
...
      "SessionEventResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "semester_id": {
            "type": "string",
            "format": "uuid",
            "title": "Semester Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "slot_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Number"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "title": "End Time"
          },
          "subject_name": {
            "type": "string",
            "title": "Subject Name"
          },
          "lesson_type": {
            "type": "string",
            "title": "Lesson Type"
          },
          "teacher_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teacher Name"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "classroom_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classroom Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "semester_id",
          "date",
          "slot_number",
          "start_time",
          "end_time",
          "subject_name",
          "lesson_type",
          "teacher_name",
          "group_name",
          "classroom_name"
        ],
        "title": "SessionEventResponse"
      },
...
      "TemplateResponse": {
        "properties": {
          "semester_id": {
            "type": "string",
            "format": "uuid",
            "title": "Semester Id"
          },
          "semester_name": {
            "type": "string",
            "title": "Semester Name"
          },
          "week_1": {
            "items": {
              "$ref": "#/components/schemas/DayTemplate"
            },
            "type": "array",
            "title": "Week 1"
          },
          "week_2": {
            "items": {
              "$ref": "#/components/schemas/DayTemplate"
            },
            "type": "array",
            "title": "Week 2"
          }
        },
        "type": "object",
        "required": ["semester_id", "semester_name", "week_1", "week_2"],
        "title": "TemplateResponse"
      },
...
      "TemplateSlotResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "semester_id": {
            "type": "string",
            "format": "uuid",
            "title": "Semester Id"
          },
          "week_number": {
            "type": "integer",
            "title": "Week Number"
          },
          "day_of_week": {
            "type": "integer",
            "title": "Day Of Week"
          },
          "slot_number": {
            "type": "integer",
            "title": "Slot Number"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "title": "End Time"
          },
          "subject_name": {
            "type": "string",
            "title": "Subject Name"
          },
          "lesson_type": {
            "type": "string",
            "title": "Lesson Type"
          },
          "teacher_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teacher Name"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "classroom_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classroom Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "semester_id",
          "week_number",
          "day_of_week",
          "slot_number",
          "start_time",
          "end_time",
          "subject_name",
          "lesson_type",
          "teacher_name",
          "group_name",
          "classroom_name"
        ],
        "title": "TemplateSlotResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}